chore: dropdown menu, edit safe button added and increased homepage PAGE_SIZE
