+++ src/pages/safe.tsx
@@ -1,4 +1,4 @@
-import { useContext } from 'react'
+import { useContext, useState } from 'react'
 import NavbarLayout from 'src/libraries/ui/navbarLayout'
 import { ApiClientsContext } from 'src/pages/_app'
 import AddToSafe from 'src/v2/components/Safe/AddToSafe'
@@ -8,8 +8,10 @@ import Dashboard from 'src/v2/components/Safe/Dashboard'
 function Safe() {
 	const { workspace } = useContext(ApiClientsContext)!
 
+	const [edit, setEdit] = useState<boolean>(false)
+
 	return (
-		workspace?.safe ? <Dashboard /> : <AddToSafe />
+		workspace?.safe && !edit ? <Dashboard setEdit={setEdit} /> : <AddToSafe />
 	)
 }
